services/renderer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.9.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
